# go/cmd/maxi-knows/main.go
package main

import (
	"crypto/md5"
	"database/sql"
	"encoding/json"
	"fmt"
	"html/template"
	"log"
	"net/http"
	"strings"

	_ "whoknows/docs"

	httpSwagger "github.com/swaggo/http-swagger"

	_ "modernc.org/sqlite"
)

// PageData contains the data that can be passed from the Go backend
// to the HTML templates.
type PageData struct {
	Username      string
	Flashes       []string
	Error         string
	FormUsername  string
	FormEmail     string
	Query         string
	SearchResults []SearchResult
}

// SearchResult represents one result shown on the search page.
type SearchResult struct {
	URL         string
	Title       string
	Description string
}

// renderTemplate handles the shared template rendering logic.
// It loads layout.html together with the page-specific template,
// then renders the named "layout" template and sends it to the browser.
func renderTemplate(w http.ResponseWriter, page string, data PageData) {
	tmpl, err := template.ParseFiles(
		"templates/layout.html",
		"templates/"+page,
	)

	// Stop and return a 500 error if the templates cannot be loaded.
	if err != nil {
		http.Error(w, "Could not load template", http.StatusInternalServerError)
		return
	}

	// Render the named "layout" template and send the generated HTML
	// to the browser through the ResponseWriter.
	err = tmpl.ExecuteTemplate(w, "layout", data)

	// Return a 500 error if the template could not be rendered.
	if err != nil {

		log.Printf("Could not render template: %v", err)

		http.Error(w, "Could not render template", http.StatusInternalServerError)
	}
}

// pageHandler creates a HTTP handler for pages that only need
// to render a template using the shared layout.
//
// The page parameter decides which page-specific template is rendered.
// For example, pageHandler("about.html") renders about.html together
// with layout.html.
func pageHandler(page string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {

		// Create the data object that is passed to the templates.
		// Username and Flashes are empty for now.
		data := PageData{}

		renderTemplate(w, page, data)
	}
}

// searchPageHandler handles GET requests to the search page.
// It reads the search query from the URL and passes it to the template.
func searchPageHandler(w http.ResponseWriter, r *http.Request) {
	data := PageData{
		Query:         r.URL.Query().Get("q"),
		SearchResults: []SearchResult{},
	}

	renderTemplate(w, "search.html", data)
}

// apiSearchHandler godoc
// @Summary      Search pages
// @Description  Search pages by query and language
// @Tags         search
// @Produce      json
// @Param        q         query     string  true   "Search query"
// @Param        language  query     string  false  "Language code" default(en)
// @Success      200  {object}  map[string]interface{}
// @Failure      422  {object}  map[string]interface{}
// @Router       /search [get]
func apiSearchHandler(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		q := r.URL.Query().Get("q")
		language := r.URL.Query().Get("language")

		if !r.URL.Query().Has("q") {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 422,
				"message":    "q is required",
			})
			return
		}

		if language == "" {
			language = "en"
		}

		rows, err := db.Query(
			"SELECT url, title, description FROM pages WHERE language = ? AND content LIKE ?",
			language, "%"+q+"%",
		)

		if err != nil {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}
		defer rows.Close()

		results := []SearchResult{}

		for rows.Next() {
			var result SearchResult

			err = rows.Scan(&result.URL, &result.Title, &result.Description)

			if err != nil {
				http.Error(w, "Database error", http.StatusInternalServerError)
				return
			}

			results = append(results, result)
		}

		if err = rows.Err(); err != nil {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)

		json.NewEncoder(w).Encode(map[string]interface{}{
			"data": results,
		})
	}
}

// apiRegisterHandler godoc
// @Summary      Register a new user
// @Description  Creates a new user account
// @Tags         auth
// @Accept       x-www-form-urlencoded
// @Produce      json
// @Param        username   formData  string  true  "Username"
// @Param        email      formData  string  true  "Email"
// @Param        password   formData  string  true  "Password"
// @Param        password2  formData  string  true  "Password confirmation"
// @Success      200  {object}  map[string]interface{}
// @Failure      422  {object}  map[string]interface{}
// @Failure      409  {object}  map[string]interface{}
// @Router       /register [post]
func apiRegisterHandler(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		r.ParseForm()

		if !r.Form.Has("username") ||
			!r.Form.Has("email") ||
			!r.Form.Has("password") ||
			!r.Form.Has("password2") {

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{})
			return
		}

		username := r.Form.Get("username")
		email := r.Form.Get("email")
		password := r.Form.Get("password")
		password2 := r.Form.Get("password2")

		//Check if username is empty
		if username == "" {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 422,
				"message":    "Username is required",
			})
			return
		}

		//Check if email is empty or invalid format
		if email == "" || !strings.Contains(email, "@") {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 422,
				"message":    "Valid email is required",
			})
			return
		}

		//check if password is empty
		if password == "" {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 422,
				"message":    "Password is required",
			})
			return
		}

		//check if passwords match
		if password != password2 {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 422,
				"message":    "Passwords do not match",
			})
			return
		}

		//check if username or email already exists in db
		var existingUserID int

		err := db.QueryRow(
			"SELECT id FROM users WHERE username = ? OR email = ?",
			username, email,
		).Scan(&existingUserID)

		//If username or email already exists, error message 409 shows
		if err == nil {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusConflict)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 409,
				"message":    "Username or email already exists",
			})
			return
		}

		if err != sql.ErrNoRows {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}

		// Hash the password using MD5
		passwordHash := fmt.Sprintf("%x", md5.Sum([]byte(password)))

		// Insert the new user into the database
		_, err = db.Exec(
			"INSERT INTO users (username, email, password) VALUES (?, ?, ?)",
			username, email, passwordHash,
		)

		if err != nil {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)

		json.NewEncoder(w).Encode(map[string]interface{}{
			"statusCode": 200,
			"message":    "Registered",
		})
	}
}

// apiLoginHandler godoc
// @Summary      Log in a user
// @Description  Authenticates a user with username and password
// @Tags         auth
// @Accept       x-www-form-urlencoded
// @Produce      json
// @Param        username  formData  string  true  "Username"
// @Param        password  formData  string  true  "Password"
// @Success      200  {object}  map[string]interface{}
// @Failure      401  {object}  map[string]interface{}
// @Failure      422  {object}  map[string]interface{}
// @Router       /login [post]
func apiLoginHandler(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		r.ParseForm()

		if !r.Form.Has("username") || !r.Form.Has("password") {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnprocessableEntity)

			json.NewEncoder(w).Encode(map[string]interface{}{})
			return
		}

		username := r.Form.Get("username")
		password := r.Form.Get("password")

		row := db.QueryRow(
			"SELECT id, username, password FROM users WHERE username = ?",
			username,
		)

		var userID int
		var storedUsername string
		var storedPassword string

		err := row.Scan(&userID, &storedUsername, &storedPassword)

		if err == sql.ErrNoRows {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnauthorized)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 401,
				"message":    "Invalid username",
			})
			return
		}

		if err != nil {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}

		passwordHash := fmt.Sprintf("%x", md5.Sum([]byte(password)))

		if storedPassword != passwordHash {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusUnauthorized)

			json.NewEncoder(w).Encode(map[string]interface{}{
				"statusCode": 401,
				"message":    "Invalid password",
			})
			return
		}

		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)

		json.NewEncoder(w).Encode(map[string]interface{}{
			"statusCode": 200,
			"message":    "Logged in",
		})
	}
}

// apiLogoutHandler godoc
// @Summary      Log out the current user
// @Description  Logs the current user out
// @Tags         auth
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Router       /logout [get]
func apiLogoutHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	json.NewEncoder(w).Encode(map[string]interface{}{
		"statusCode": 200,
		"message":    "Logged out",
	})
}

// @title           WhoKnows API
// @version         1.0
// @description     Search, register, login and logout API for WhoKnows.
// @host            localhost:8080
// @BasePath        /api
func main() {
	db, err := sql.Open("sqlite", "../data/whoknows.db")
	if err != nil {
		log.Fatal(err)
	}
	defer db.Close()

	// HTML routes

	// GET /SEARCH
	http.HandleFunc("GET /{$}", searchPageHandler)

	// GET /REGISTER
	http.HandleFunc("GET /register", pageHandler("register.html"))

	// GET /LOGIN
	http.HandleFunc("GET /login", pageHandler("login.html"))

	// GET /ABOUT
	// Render the about page using the shared pageHandler.
	http.HandleFunc("GET /about", pageHandler("about.html"))

	// Serve files from the static folder.
	// For example:
	// /static/style.css -> static/style.css
	// /static/monkgroup.png -> static/monkgroup.png
	http.Handle(
		"/static/",
		http.StripPrefix("/static/", http.FileServer(http.Dir("static"))),
	)

	// ============================ API routes ============================
	// GET API/SEARCH
	http.HandleFunc("GET /api/search", apiSearchHandler(db))

	// POST /API/REGISTER
	http.HandleFunc("POST /api/register", apiRegisterHandler(db))

	// POST /API/LOGIN
	http.HandleFunc("POST /api/login", apiLoginHandler(db))

	// GET /API/LOGOUT
	http.HandleFunc("GET /api/logout", apiLogoutHandler)

	// Swagger UI + JSON spec
	http.Handle("/swagger/", httpSwagger.WrapHandler)

	// Start the web server on port 8080.
	log.Println("Server running on http://localhost:8080")
	log.Fatal(http.ListenAndServe(":8080", nil))
}